[Autofill Log] : log details about views filling
Add logs for tracking failures and success on views filling.
Note that for the success cases, these are the upper bound. It's
calculated when the framework sends the autofill values to the view.
It's possible that the views reject autofill for some reason, and
doesn't autofill. Since autofill doesn't have any way to know that,
we log what we can.

Also, this CL fixes cases where multiple PresentationEventLogs may be
generated for a single presentation.

Test: Manually verified events being generated, and
      atest CtsAutoFillServiceTestCases (WIP)
Bug: 319872477
Change-Id: I1de1ec7518d0fda9c978b47396a6aac4ed347175