Allowing Private Space unlock without device lock.
Currently, Private profile unlock is disabled if
device lock is not present.
However, private space is still allowed to be created
without device lock via adb commands.

In this change we allow the unlock if the space was created
via adb (by checking user setup complete), when device lock
is not present.

As user setup is always marked complete when PS is created via
Settings entry-point, this change does not effect regular flows.

Bug: 322882655
Test: Manual build and flash
Change-Id: I7e1a3ecd63d78fb3dcca04a78923f98289eb8bf3