VisualInterruptionDecisionProviderTestBase: test app suspended
If an app is suspended in focus mode, we should suppress peek and pulse;
otherwise, the notifications may *immediately* peek or pulse when focus
mode ends.

Add tests that we suppress peek and pulse while an app is suspended.

These will fail, but the other two changes in this topic will implement
the change and make the tests pass.

Note that, despite touching code in the visual interruptions refactor,
this change is *not* related to or flagged by the refactor's feature
flag; this is a long-standing omission in our interruption logic, not a
bug in the refactored code.

Bug: 329620290
Flag: NA
Test: atest NotificationInterruptStateProviderWrapperTest
Test: atest VisualInterruptionDecisionProviderImplTest
Change-Id: Ia94dbde4f112595a4ec5a30800b9f5a2e6437fcc