Improve performance of VRR
Bug: 330765659

Trace.isTagEnabled() is expensive and has been removed in favor
of the less expensive condition it was protecting.

Size checks are done only when the size of a View changes.
The small size requirement is calculated once rather than
in every View.

For Views that don't draw, only vote when motion is involved.

Test: ran performance check. Improvement of around 30%

Change-Id: I73658e94d41799a7d88fcc568061f8d46c90c757