Moving hardware layer to the parent of brightness slider
Hardware layer was introduced to make alpha animation better - without it, slider_foreground (thin line in the middle of slider) was alpha-animating at the same time as rest of slider and this caused that thin line to be visible through slider while fading out. Hardware layer causes the whole view texture to be animated, instead of each element separately.
We still need it but for incoming focus state of brightness slider we also want to draw outside of the BrightnessSliderView bounds - hardware layer prevents us from doing this.
Moving hardware layer to QSContainerImpl because its changing alpha is animating whole QS panel in split shade and it's the lowest parent that wouldn't clip focus outline.

Bug: 329244723
Flag: None
Test: Expand/collapse split shade, in portrait expand/collapse QS to QQS and see brightness slider animates nicely
Change-Id: If730e7a04eb1b9f6524ad5c1e6a616c0a48944e7