Fixed graphical glitch in lockpattern view
Fixed end point correctional animation for cell to cell animations
during lock pattern drawing.

In the lock pattern view, cells have an area of slop around them to
allow for drawing the pattern without hitting the exact center of cells.

When a cell is hit, the end point of the line drawn between that cell
and the previous is adjusted with an animation from the hit coordinate
and the exact center of the cell.

With LockPatternView_keepDotActivated set to true, cell animations for
the previous cell are recreated when a new cell is hit to allow changing
how a cells color is updated as the pattern is drawn.

This caused a glitch where the starting position of the end point
correcting animation for said previous cell would be recreated with the
current finger position coordinate, making it jump ahead to be closer to
the current latest cell, and animating back towards the target.

This fix makes the animation instead continue from where it was last, in
the case where the animation is recreated.

Test: manual, set a short, tight lock pattern, and try it in rapid
motion.
Bug: 328088937

Change-Id: Id445266cbc834fec7a514dbb227557f2ceae9dba