Skip setting intermediate top when moving task to front
For example:
Start a new activity on a background task with an existing activity.
When moving the task to front, TDA#positionChildTaskAt ->
updateTopResumedActivityIfNeeded can be called before the new activity
is added on the task. Then the existing non-visible activity will be
set as top app within a short time. That will cause some resource
managements confusion about process importance.

Bug: 335085960
Test: ActivityTaskSupervisorTests#testUpdateTopResumed_moveToFront
      ResourceManagerMultiTest#testReclaimResource
Change-Id: I154b3afdff978ffceeea2dd34fa467baa20aa015