Revert^2 "[Sat] Add satellite supported changed callback"
This reverts commit 0bc671f1ff249a5fb19c15c5e98533a708918d3e.

Reason for revert: Added a a try/catch to handle the exception

Original commit message:

Per the linked bug, there is a new API on `SatelliteManager` that
implements a callback for satellite support. This is in addition to the
existing query API that happens when the repository starts up.

This CL adds the listener after the initial query, and makes sure that
if satellite support shows up later it is caught. The state change
listener runs regardless of the initial supported state.

Test: DeviceBasedSatelliteRepositoryImplTest
Bug: 342278770
Flag: NONE bugfix
Change-Id: I2ddf754b60ca71427754c3940f593e851c9c28da