Reland "Fix ShadeTouchHandler over the lock screen"
The previous CL changed where ShadeTouchHandler sends touches, which
broke opening the shade over the dream on AP3A builds, which don't have
the hub or some of the keyguard refactors.

The fix is to only send touches on the new path for builds with the
hub flag enabled, and when not dreaming. The latter logic change is new
and added to prevent shade flickering when opening and closing the hub
over the dream.

To summarize, ShadeTouchHandler will continue to send touches to the
ShadeViewController when dreaming on all builds, and will send touches
to NotificationShadeWindowViewController only when on the glanceable
hub over the lock screen.

This reverts commit fa16e7d0b9b12d5ee308021810a24ecaa11f4a33.

Bug: 328838259
Fix: 328838259
Test: atest ShadeTouchHandlerTest NotificationShadeWindowViewControllerTest
      Also verified manually that opening shade works on dream, lock
      screen, hub over either surface, and dream on AP3A build
Flag: com.android.systemui.glanceable_hub
Change-Id: I4fed54921d69606f4224505713883d1a186c7f49