Update shade transition on full expand or collapse
On the lockscreen, this is normally done by a call in
LockscreenShadeTransitionController when updating dragDownAmount.
However it seems that sometimes this update doesn't happen, leaving
MediaHierarchyManager in a bad state. This change listens to the
ShadeInteractor as a backup to ensure the state is updated when the
transition has finished.

Bug: 319244625
Flag: com.android.systemui.media_controls_lockscreen_shade_bug_fix
Test: manual - swipe shade while locked and unlocked and landscape mode
Test: manual - same on tablet device
Change-Id: I5c310ab038603c33f0a6a202ebcee9110e0612d9