A11yMenuService unregisters displayListener
The DisplayListener wasn't being unregistered properly,
which seems to have led to cases in which
it can be called while mA11yMenuLayout is null,
causing a crash.

This change unregisteres the listener properly,
and does a safety null check as well.

(This change also fixes the indentation level of the listener)

Test: N/A
Flag: EXEMPT small internal change
Bug: 346305743
Change-Id: Ic89bbf91f0d615898b7c96a6bb38b5673da40eed